+++ src/matcher.ts
@@ -12,36 +12,43 @@ import {KeysOf, TypeExt, Variant, VariantCreator, VariantsOfUnion} from './varia
  * Finish handling with a terminal: `.complete()`, `.execute()`, or `.else()`.
  */
 export type Matcher<
-    T extends TypeExt<'type', string>,
-    H extends Partial<Handler<VariantsOfUnion<T>>>,
+    T extends TypeExt<K, string>,
+    H extends Partial<Handler<VariantsOfUnion<T, K>>>,
     K extends string = 'type',
 > = {
+    /**
+     * The in-progress handler.
+     */
     readonly handler?: H;
     /**
-     * The target object being matched
+     * The target object being matched.
      */
     readonly target: T;
     /**
+     * The key used as the discriminant.
+     */
+    readonly key: K;
+    /**
      * Handle one or more cases using a `match`-like handler object.
      */
-    when<HPrime extends SplayPartial<Exclude<T, TypeExt<'type', keyof H>>>>(hp: HPrime): Matcher<T, H & HPrime, K>;
+    when<HPrime extends SplayPartial<Exclude<T, TypeExt<K, keyof H>>, K>>(hp: HPrime): Matcher<T, H & HPrime, K>;
     /**
      * Handle one or more cases by specifying the relevant types in an array
      * then providing a function to process a variant of one of those types.
      */
     when<
-        KPrime extends (Exclude<T['type'], keyof H> | VariantCreator<Exclude<T['type'], keyof H>, Func>),
-        HFunc extends (x: ExtractOfUnion<T, GetKey<KPrime>>) => any,
+        KPrime extends (Exclude<T[K], keyof H> | VariantCreator<Exclude<T[K], keyof H>, Func, K>),
+        HFunc extends (x: ExtractOfUnion<T, GetKey<KPrime, K>, K>) => any,
     >(
         keys: KPrime[],
         handler: HFunc,
-    ): Matcher<T, H & Record<GetKey<KPrime>, HFunc>, K>;
+    ): Matcher<T, H & Record<GetKey<KPrime, K>, HFunc>, K>;
     /**
      * Execute the match immediately. Exhaustiveness is not guaranteed, but if the matcher
      * has missing cases this function may return `undefined`. In that scenario the `undefined` 
      * type will be added to this function's possible return types.
      */
-    execute(): Exclude<T['type'], keyof H> extends never 
+    execute(): Exclude<T[K], keyof H> extends never 
         ? ReturnType<EnsureFunc<H[keyof H]>> 
         : (ReturnType<EnsureFunc<H[keyof H]>> | undefined)
     ;
@@ -49,9 +56,9 @@ export type Matcher<
      * Take any remaining cases and handle them in a single function. This is a terminal and
      * will execute the match immediately. If a previous `when()` statement 
      */
-    else<F extends (variant: Exclude<T, TypeExt<'type', keyof H>>) => any>(f: F): ReturnType<EnsureFunc<H[keyof H]> | F>;
+    else<F extends (variant: Exclude<T, TypeExt<K, keyof H>>) => any>(f: F): ReturnType<EnsureFunc<H[keyof H]> | F>;
 
-} & (Exclude<T['type'], keyof H> extends never ? {
+} & (Exclude<T[K], keyof H> extends never ? {
     /**
      * Only exists if all cases are handled.
      */
@@ -59,7 +66,7 @@ export type Matcher<
 } : {});
 
 
-type GetKey<T extends (string | VariantCreator<string, Func>)> = T extends string ? T : T extends VariantCreator<infer TT, Func> ? TT : never;
+type GetKey<T extends (string | VariantCreator<string, Func, K>), K extends string = 'type'> = T extends string ? T : T extends VariantCreator<infer TT, Func, K> ? TT : never;
 
 
 type Coalesce<T, U> = T extends undefined ? U : T;
@@ -72,18 +79,20 @@ type EnsureFunc<T> = T extends Func ? T : never;
  * @param _typeKey 
  */
 export function matcher<
-    T extends TypeExt<'type', string>,
+    T extends TypeExt<K, string>,
     H extends {},
     K extends string = 'type',
 >(target: T, handler = {} as H, _typeKey = 'type' as K): Matcher<T, {}, K> {
     return {
         target,
         handler,
-        when: function<HPrime extends SplayPartial<Exclude<T, TypeExt<'type', keyof H>>>>(
+        key: _typeKey ?? 'type',
+        when: function<HPrime extends SplayPartial<Exclude<T, TypeExt<K, keyof H>>, K>>(
             this: Matcher<T, H, K>, 
-            hp: HPrime | Array<T['type'] | VariantCreator<T['type']>>,
+            hp: HPrime | Array<T[K] | VariantCreator<T[K]>>,
             hfunc?: Func
         ) {
+            // The case where an array is passed in.
             if (Array.isArray(hp)) {
                 const keys = hp.map(keyOrVC => {
                     if (typeof keyOrVC === 'string') {
@@ -120,14 +129,14 @@ export function matcher<
         } as Func,
         execute: function(this: Matcher<T, H, K>) {
             console.log(this.handler);
-            const result: (x: T) => unknown = (this.handler as any)[target.type]
-            return result(target);
+            const result: (x: T) => unknown = (this.handler as any)[this.target[this.key]]
+            return result(this.target);
         } as Func,
         complete: function(this: Matcher<T, H, K>) {
             return this.execute();
         } as Func,
         else: function(this: Matcher<T, H, K>, func: (x: T) => unknown) {
-            if (this.target.type in (this.handler ?? {})) {
+            if (this.target[this.key] in (this.handler ?? {})) {
                 return this.execute();
             } else {
                 return func(this.target);
@@ -138,16 +147,16 @@ export function matcher<
 
 
 
-type SplayPartial<T extends TypeExt<'type', string>> =
-    Identity<SplayCase<T, T['type']>[T['type']]>;
+type SplayPartial<T extends TypeExt<K, string>, K extends string = 'type'> =
+    Identity<SplayCase<T, K>[T[K]]>;
 
-type SplayCase<T extends TypeExt<'type', Types>, Types extends string> = {
-    [P in T['type']]: KeyCase<T, T['type'], P, (x: ExtractOfUnion<T, P, 'type'>) => any>;
+type SplayCase<T extends TypeExt<K, string>,K extends string = 'type'> = {
+    [P in T[K]]: KeyCase<T, T[K], P, (x: ExtractOfUnion<T, P, K>) => any, K>;
 }
 
 /**
  * 
  */
-type KeyCase<T extends TypeExt<'type', Types>, Types extends string, Type extends Types, F extends (...args: any[]) => any> 
-    = Partial<Handler<VariantsOfUnion<T>>> & TypeExt<Type, F>;
+type KeyCase<T extends TypeExt<K, string>, Types extends string, Type extends Types, F extends (...args: any[]) => any, K extends string = 'type'> 
+    = Partial<Handler<VariantsOfUnion<T, K>>> & TypeExt<Type, F>;