+++ src/tools.ts
@@ -1,4 +1,4 @@
-import variant, {Outputs, Variant, VariantCreator, WithProperty} from "./variant";
+import variant, {Outputs, Property, Variant, VariantCreator} from "./variant";
 import {ExtractOfUnion, Func, Identity} from "./util";
 
 /**
@@ -69,7 +69,7 @@ export function payload<T>(_example?: T) {
 }
 
 export function property<K extends string = 'payload'>(key: K) {
-    return <T> () => (payload: T) => ({[key]: payload}) as K extends keyof infer KLiteral ? WithProperty<keyof KLiteral & string, T> : never;
+    return <T> () => (payload: T) => ({[key]: payload}) as K extends keyof infer KLiteral ? Property<keyof KLiteral & string, T> : never;
 }
 
 export function data<T>(x: T) {
@@ -91,7 +91,7 @@ type FlattenToTypeStr<T extends (string | VariantCreator<string, Func, K>), K ex
  * Curried isType, useful for `.filter` or rxjs
  * @param type 
  */
-export function isType<T extends (string | VariantCreator<string, Func, K>), K extends string = 'type'>(type: T): <O extends WithProperty<K, string>> (o: O) => o is ExtractOfUnion<O, FlattenToTypeStr<T, K>, K>; 
+export function isType<T extends (string | VariantCreator<string, Func, K>), K extends string = 'type'>(type: T): <O extends Property<K, string>> (o: O) => o is ExtractOfUnion<O, FlattenToTypeStr<T, K>, K>; 
 /**
  * Check if an object is of a given type. The type here
  * may be a string or a variant constructor (i.e. `Animal.dog`).
@@ -105,7 +105,7 @@ export function isType<T extends (string | VariantCreator<string, Func, K>), K e
  * @param key optional discriminant key override. 'type' by default.
  */
 export function isType<
-    O extends WithProperty<K, string>,
+    O extends Property<K, string>,
     T extends (O[K] | VariantCreator<O[K], Func, K>),
     K extends string = 'type',
 >(
@@ -114,7 +114,7 @@ export function isType<
     key?: K,
 ): instance is ExtractOfUnion<O, T extends VariantCreator<infer R, Func, K> ? R : T extends string ? T : never, K>; 
 export function isType<
-    O extends WithProperty<K, string>,
+    O extends Property<K, string>,
     T extends (O[K] | VariantCreator<O[K], Func, K>),
     K extends string = 'type',
 >(
@@ -126,13 +126,13 @@ export function isType<
         if (typeof instanceOrType === 'function' || typeof instanceOrType === 'string') {
             const typeArg = instanceOrType as T;
             const typeStr = typeof typeArg === 'string' ? typeArg : (typeArg as VariantCreator<string, any, K>).type;
-            return <O extends WithProperty<K, string>>(o: O): o is ExtractOfUnion<O, FlattenToTypeStr<T, K>, K> => isType(o, typeStr);
+            return <O extends Property<K, string>>(o: O): o is ExtractOfUnion<O, FlattenToTypeStr<T, K>, K> => isType(o, typeStr);
         } else {
             const instance = instanceOrType as O;
             const type = typeOrKey as T;
 
             const typeStr = typeof type === 'string' ? type : (type as VariantCreator<string, any, K>).type;
-            return instance != undefined && (instance as WithProperty<K, string>)[key ?? 'type'] === typeStr;
+            return instance != undefined && (instance as Property<K, string>)[key ?? 'type'] === typeStr;
         }
     } else {
         return false;